+++ src/district_energy_model/input_files/inputs.py
@@ -241,6 +241,19 @@ scen_techs = {
         'export_subsidy': 0 if force_asynchronous_storage else no_force_asynchronous_storage_export_subsidy,
         'only_use_installed':False,
         },
+
+    'solar_pvalpine':{ # pv
+        'deployment':True,
+        'kWp_max':'inf', 
+        'eta_overall': -1.0, #Not used by code.
+        'lifetime':25,
+        'base_capex':6000, #capex per installed kWp power. This is not the same as the peak of the solar PV production.
+        'base_maintenance_cost': 13.0,
+        'interest_rate':interest_rate,
+        'export_subsidy': 0 if force_asynchronous_storage else no_force_asynchronous_storage_export_subsidy,
+        'only_use_installed':False,
+        },
+
         
     'solarthermal_rooftop':{ # solar
         'deployment':True,